refactor: correct card list UI ownership
