+++ src/features/card/view/index.ts
@@ -1,0 +1,3 @@
+export { BackText, type BackTextProps } from "./ui/BackText";
+export { CardOverlay } from "./ui/CardOverlay";
+export { FrontText } from "./ui/FrontText";
