+++ src/features/card/view/ui/FrontText.spec.tsx
